Revert "gtaxl: workaround for offline charging"
This reverts commit 667ef4c515ec8256e84abb7eb7f04188899527d9.

This workaround means that there is a sysmmu page read fault that causes
a full reboot when trying to charge offline. If this is intentional, this
always assumes that the device has enough power to boot, which makes this
a bad workaround. Now that the property which enabled booting straight
from offline charging is removed and so it's disabled, no workaround is
necessary as the issues were caused by that feature and no issues exist
anymore after disabling it.

Change-Id: I45af1d120d250332d6fa74022324a06a2035767f